+++ src/Delivery/ImageFile.php
@@ -1,10 +1,11 @@
 <?php
 /**
- * @copyright 2015 Contentful GmbH
+ * @copyright 2015-2017 Contentful GmbH
  * @license   MIT
  */
 
 namespace Contentful\Delivery;
+use Contentful\ImageOptions;
 
 class ImageFile extends File
 {