+++ src/Delivery/Client.php
@@ -24,7 +24,7 @@ use Contentful\Link;
  */
 class Client extends BaseClient
 {
-    const VERSION = '0.8.0-beta';
+    const VERSION = '0.9.0-dev';
 
     /**
      * @var ResourceBuilder