+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased](https://github.com/contentful/ContentfulBundle/compare/0.7.0-beta...HEAD)
+
 ## [0.7.0-beta](https://github.com/contentful/contentful.php/tree/0.7.0-beta) (2017-04-06)
 
 **ATTENTION**: This release contains breaking changes. Please take extra care when updating to this version.