+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 * Removed all get* methods except `getQueryData()` and `getQueryString()` from the various query classes. **[BREAKING]**
 
+### Fixed
+* The `FilesystemCache` would try to read cached content types from the wrong file name.
+
 ## [0.7.0-beta](https://github.com/contentful/contentful.php/tree/0.7.0-beta) (2017-04-06)
 
 **ATTENTION**: This release contains breaking changes. Please take extra care when updating to this version.