+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * The `FilesystemCache` would try to read cached content types from the wrong file name.
 * `CacheWarmer` wrote incorrect data for content types.
 * Retrieving a cached content type would cause the maximum function nesting level to be exceeded.
+* Correctly set the `Accept` header for API versioning. Previously the `Content-Type` header was set instead.
 
 ## [0.7.0-beta](https://github.com/contentful/contentful.php/tree/0.7.0-beta) (2017-04-06)